+++ www/modules/email/AccountDialog.js
@@ -164,7 +164,11 @@ GO.email.AccountDialog = function(config) {
 			fieldLabel : t("Username"),
 			name : 'username',
 			allowBlank : false,
+			autocomplete: "new-password",
 			listeners : {
+				render : function(f) {
+					f.getEl().dom.setAttribute("data-lpignore", "true");
+				},
 				change : function() {
 					this.refreshNeeded = true;
 				},
@@ -180,10 +184,14 @@ GO.email.AccountDialog = function(config) {
 		}),
 		this.ImapPasswordField = new Ext.form.TextField({
 			autocomplete: 'new-password',
+
 			fieldLabel : t("Password"),
 			name : 'password',
 			inputType : 'password',
 			listeners : {
+				render : function(f) {
+					f.getEl().dom.setAttribute("data-lpignore", "true");
+				},
 				change : function() {
 					this.refreshNeeded = true;
 				},
@@ -337,9 +345,15 @@ GO.email.AccountDialog = function(config) {
 				scope:this
 			}
 		}),this.smtpUsername= new Ext.form.TextField({
-			fieldLabel : t("Username"),
-			name : 'smtp_username',
-			disabled:true
+				fieldLabel : t("Username"),
+				name : 'smtp_username',
+				disabled:true,
+				autocomplete: "new-password",
+				listeners: {
+					render : function(f) {
+						f.getEl().dom.setAttribute("data-lpignore", "true");
+					}
+				}
 		}),
 		this.smtpPasswordStore = new Ext.ux.form.XCheckbox({
 			checked: true,
@@ -352,7 +366,12 @@ GO.email.AccountDialog = function(config) {
 			fieldLabel : t("Password"),
 			name : 'smtp_password',
 			inputType : 'password',
-			disabled:true
+			disabled:true,
+			listeners: {
+				render : function(f) {
+					f.getEl().dom.setAttribute("data-lpignore", "true");
+				}
+			}
 		})]
 	});