+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Core: totals input wider to fit 4 digits. Fixes #843
 - Core: no timeout on upgrade and cron lock
 - Calendar: Wrong calendar list when copying an event after the first time
+- Email: Add data-lpignore=true to username and password fields to disable lastpass autofill
 
 18-07-2022 6.6.103
 - Billing: Fix overwrite uploaded logo for PDF template